edge_gating

+++ model.py
@@ -279,7 +279,7 @@ class SummarizationModel(object):
         for b in range(batch_size):
             for l in range(max_labels):
                 t_indices = tf.transpose(adj_in[b][l].indices, [1, 0])
-                t_indices += max_words * l
+                t_indices += max_words * tf.cast(l, dtype=tf.int64)
                 indices.append(t_indices)
                 b_data.append(tf.ones([indices[b].shape[0]]) * l)
         indices = tf.stack(indices, axis=0)
